# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContractAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type UserPlan = (typeof UserPlan)[keyof typeof UserPlan];

export const UserPlan = {
  free: "free",
  pro: "pro",
  premium: "premium",
} as const;

export interface User {
  id: string;
  email: string;
  plan: UserPlan;
  contractsUsed: number;
  contractsLimit: number;
  createdAt?: string;
}

export type ContractStatus =
  (typeof ContractStatus)[keyof typeof ContractStatus];

export const ContractStatus = {
  uploaded: "uploaded",
  extracting: "extracting",
  extracted: "extracted",
  analyzing: "analyzing",
  analyzed: "analyzed",
  failed: "failed",
} as const;

export interface Contract {
  id: string;
  userId: string;
  filename: string;
  fileSize: number;
  status: ContractStatus;
  createdAt: string;
  analyzedAt?: string | null;
}

export type AnalysisRiskLevel =
  (typeof AnalysisRiskLevel)[keyof typeof AnalysisRiskLevel];

export const AnalysisRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Analysis {
  id: string;
  contractId: string;
  summary: string;
  risks: string[];
  keyClauses: string[];
  riskLevel: AnalysisRiskLevel;
  createdAt: string;
}

export type ContractWithAnalysis = Contract & {
  analysis?: Analysis | null;
  extractedText?: string | null;
};

export interface ChatRequest {
  message: string;
}

export interface ChatResponse {
  reply: string;
  messageId: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: string;
  contractId: string;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export type CheckoutRequestPlan =
  (typeof CheckoutRequestPlan)[keyof typeof CheckoutRequestPlan];

export const CheckoutRequestPlan = {
  pro: "pro",
  premium: "premium",
} as const;

export interface CheckoutRequest {
  plan: CheckoutRequestPlan;
}

export interface CheckoutResponse {
  success: boolean;
  checkout_url: string;
}

export type DashboardStatsPlan =
  (typeof DashboardStatsPlan)[keyof typeof DashboardStatsPlan];

export const DashboardStatsPlan = {
  free: "free",
  pro: "pro",
  premium: "premium",
} as const;

export interface DashboardStats {
  totalContracts: number;
  analyzedContracts: number;
  highRiskContracts: number;
  planUsagePercent: number;
  plan: DashboardStatsPlan;
  contractsUsed: number;
  contractsLimit: number;
}

export interface RecentActivity {
  id: string;
  contractId: string;
  filename: string;
  action: string;
  riskLevel?: string | null;
  createdAt: string;
}

export type UploadContractBody = {
  file: Blob;
};