# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ContractAI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: contracts
    description: Contract management
  - name: analysis
    description: AI analysis
  - name: chat
    description: AI chat feature
  - name: payments
    description: Payment operations
  - name: dashboard
    description: Dashboard data

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all contracts for current user
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contracts/upload:
    post:
      operationId: uploadContract
      tags: [contracts]
      summary: Upload a contract PDF
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: Contract uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a specific contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contract details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractWithAnalysis"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contracts/{id}/analyze:
    post:
      operationId: analyzeContract
      tags: [analysis]
      summary: Analyze a contract with AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (plan limit reached)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chat/{contractId}:
    post:
      operationId: chatWithContract
      tags: [chat]
      summary: Chat with AI about a contract (premium only)
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Premium plan required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for a contract
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/checkout:
    post:
      operationId: createCheckout
      tags: [payments]
      summary: Create a checkout session for plan upgrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/webhook:
    post:
      operationId: paymentWebhook
      tags: [payments]
      summary: Handle Dodo payment webhooks
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics for current user
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/recent:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent contract activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentActivity"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        plan:
          type: string
          enum: [free, pro, premium]
        contractsUsed:
          type: integer
        contractsLimit:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, email, plan, contractsUsed, contractsLimit]

    Contract:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        filename:
          type: string
        fileSize:
          type: integer
        status:
          type: string
          enum: [uploaded, extracting, extracted, analyzing, analyzed, failed]
        createdAt:
          type: string
          format: date-time
        analyzedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, filename, fileSize, status, createdAt]

    Analysis:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        summary:
          type: string
        risks:
          type: array
          items:
            type: string
        keyClauses:
          type: array
          items:
            type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
      required: [id, contractId, summary, risks, keyClauses, riskLevel, createdAt]

    ContractWithAnalysis:
      allOf:
        - $ref: "#/components/schemas/Contract"
        - type: object
          properties:
            analysis:
              $ref: "#/components/schemas/Analysis"
              nullable: true
            extractedText:
              type: string
              nullable: true

    ChatRequest:
      type: object
      properties:
        message:
          type: string
      required: [message]

    ChatResponse:
      type: object
      properties:
        reply:
          type: string
        messageId:
          type: string
      required: [reply, messageId]

    ChatMessage:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, contractId, role, content, createdAt]

    CheckoutRequest:
      type: object
      properties:
        plan:
          type: string
          enum: [pro, premium]
      required: [plan]

    CheckoutResponse:
      type: object
      properties:
        success:
          type: boolean
        checkout_url:
          type: string
      required: [success, checkout_url]

    DashboardStats:
      type: object
      properties:
        totalContracts:
          type: integer
        analyzedContracts:
          type: integer
        highRiskContracts:
          type: integer
        planUsagePercent:
          type: number
        plan:
          type: string
          enum: [free, pro, premium]
        contractsUsed:
          type: integer
        contractsLimit:
          type: integer
      required: [totalContracts, analyzedContracts, highRiskContracts, planUsagePercent, plan, contractsUsed, contractsLimit]

    RecentActivity:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        filename:
          type: string
        action:
          type: string
        riskLevel:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, contractId, filename, action, createdAt]