# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContractAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisRiskLevel } from "./analysisRiskLevel";

export interface Analysis {
  id: string;
  contractId: string;
  summary: string;
  risks: string[];
  keyClauses: string[];
  riskLevel: AnalysisRiskLevel;
  createdAt: Date;
}