# lib/api-zod/src/generated/types/contractWithAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContractAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Analysis } from "./analysis";
import type { Contract } from "./contract";

export type ContractWithAnalysis = Contract & {
  analysis?: Analysis | null;
  extractedText?: string | null;
};