# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContractAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from "./userPlan";

export interface User {
  id: string;
  email: string;
  plan: UserPlan;
  contractsUsed: number;
  contractsLimit: number;
  createdAt?: Date;
}